requirements.txt
pytelegrambotapi
flask
pillow
python-dotenv